requirements.txt
flask==3.0.3
python-dotenv==1.0.0
ffmpeg-python==0.2.0
flask-cors==4.0.0
requests==2.32.3 